logisticInvoiceanalyser-railway-ready (1)/logisticInvoiceanalyser-fixed/backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2
pydantic==2.12.5
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.2.1
python-multipart==0.0.22
pdfplumber==0.11.9
pdfminer.six==20251230
pillow==12.1.1
openai==1.99.0
httpx==0.28.1
aiohttp==3.13.3
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pandas==3.0.1
numpy==2.4.2
